Modifications needed to permit arbitrary Civet file-naming
In the past, the Civet output needed to be named according to the scanId-scanDate format. This change introduced a new setting CIVET_SCANID_APPEND_SCANDATE_TO_KEYNAME, which can be set to ON/OFF, which permits processing of Civet scans that do not match the aforementioned naming convention.

Also -- added code to the cortical thickness processing stream to add mean thickness values by hemisphere (left, right, both), in addition to values by ROI.